requirements.txt
Flask==3.0.3
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
